+++ backend/routes/company-routes.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { getCompanyData, addCompany, updateById, getById, deleteById } = require('../controllers/company-controller');
+const { getCompanyData, addCompany, updateById, getById, deleteById, getByUserId } = require('../controllers/company-controller');
 
 
 
@@ -9,6 +9,7 @@ comRouter.get("/", getCompanyData);
 comRouter.post("/addCompany", addCompany);
 comRouter.put("/update/:id", updateById)
 comRouter.get("/:id", getById);
+comRouter.get("/user/:id", getByUserId);
 comRouter.delete("/delete/:id", deleteById)
